___Latest PLC backup
It includes: 1. PLC control of stitching 2. Full PLC control of robot programs 3. Read robot/stitching data and push to GRAFANA
___Latest Config JSON file
Includes the setup of the interface to GRAFANA. It points to the PLC OPC-UA server and gets the data for selected tags that are displayed on GRAFANA
___Latest Docker compose files
Include the setup of Docker to be able to push the data to GRAFANA from the OPC-UA server
___Latest RoboDK simulation
Includes the simulation of picking up a piece, stitching all 4 sides and getting home
___Latest RoboDK CSV custom post processor
Includes a custom CSV post processor that exports RoboDK movements/robot positions and saves them as a CSV file that the PLC can read and use to execute a robot program.